fix[cache]: Unified refinery postprocessing and data fetching by removing duplicates.
